requirements.txt
bs4==0.0.1
langchain==0.0.272
requests==2.28.2
xmltodict==0.13.0
openai==0.27.0
tiktoken==0.3.0
faiss-cpu==1.7.3
pyngrok==4.1.1
flask_ngrok==0.0.25
flask_cors==3.0.10
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
pandas
pandas-datareader==0.10.0
pandas-gbq==0.17.9
pandas-profiling==3.2.0
plotly==5.13.1
pytz
sklearn-pandas==2.2.0
typing-inspect==0.8.0
typing_extensions==4.5.0
authlib==1.2.0 
validators==0.20.0 
waitress==2.1.2
python-dotenv
pymilvus
grpcio==1.56.0
flask_socketio
